move point and transform into geometry module
