+++ docs/source/notebooks/geometry/leg.py
@@ -18,43 +18,26 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
-from pathlib import Path
-
-from ipywidgets import interact
-from matplotlib import pyplot as plt
-from matplotlib.animation import FFMpegWriter, FuncAnimation
-
-from .utils import is_sphinx_build, is_sphinx_build_no_videos
+from line import Line3D
 
 
-def animate_plot(
-    _fig,
-    _animate: callable,
-    _frames,
-    _interval=16,  # 60 fps
-    _interactive=False,
-    _save_animation_name=None,
-    **interact_kwargs,
-):
-    if is_sphinx_build_no_videos():
-        return
+class Leg3D:
+    """A 3D leg class."""
 
-    anim = None
+    def __init__(self, lines: list[Line3D]):
+        self.lines = lines
 
-    plt.rcParams['animation.html'] = 'html5'
+    def __iter__(self):
+        return iter(self.lines)
 
-    if _interactive and not is_sphinx_build():
-        with plt.ion():
-            anim = interact(_animate, frame=(0, _frames), **interact_kwargs)
-            plt.show()
-    else:
-        with plt.ioff():
-            anim = FuncAnimation(_fig, _animate, frames=_frames, interval=_interval)
-            display(anim)  # type: ignore # noqa: F821
-            plt.close(_fig)
+    @property
+    def xy(self):
+        return [line.xy for line in self.lines]
 
-            if _save_animation_name is not None:
-                animation_writer = FFMpegWriter(fps=24)
-                anim.save(Path(_save_animation_name).with_suffix('.mp4'), writer=animation_writer)
+    @property
+    def xz(self):
+        return [line.xz for line in self.lines]
 
-    return anim
+    @property
+    def yz(self):
+        return [line.yz for line in self.lines]
