+++ docs/source/notebooks/plotting/leg.py
@@ -20,10 +20,10 @@
 
 from typing import Literal
 
+from geometry import Leg3D, Line, Line3D, Point
 from inline_labels import add_inline_labels
 import matplotlib.pyplot as plt
 import numpy as np
-from point import Leg3D, Line, Line3D, Point
 
 from .angle_annotation import AngleAnnotation
